% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdGrad.R
\name{fdGrad}
\alias{fdGrad}
\title{Finite-difference gradient}
\usage{
fdGrad(pars, fun, ..., .relStep = (.Machine$double.eps)^(1/2), minAbsPar = 0)
}
\arguments{
\item{pars}{Parameters involved in gradient computation}

\item{fun}{Function for which the finite difference gradient is computed}

\item{...}{Other arguments for \code{fun} not involved in gradient computation}

\item{.relStep}{Amount to shift parameter values when calculating gradient}

\item{minAbsPar}{Minimum absolute parameter value}
}
\value{
Finite difference gradient
}
\description{
Finite-difference gradient
}
