% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_posterior.R
\name{log_posterior}
\alias{log_posterior}
\title{Log posterior for PK parameters}
\usage{
log_posterior(lpr, ivt, dat)
}
\arguments{
\item{lpr}{log-PK parameter vector}

\item{ivt}{List with containing start of infusion times, end of infusion times,
and rate of infusion at each dose}

\item{dat}{Concentration data frame of the form: data.frame(time_h, conc_mg_dl)}
}
\value{
The log-posterior distribution evaluated at the specified log-parameter vector
}
\description{
Log posterior for PK parameters
}
