% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mic_stat.R
\name{mic_stat}
\alias{mic_stat}
\title{Time Above Pharmacokinetic Threshold}
\usage{
mic_stat(pk_pars, ivt, tms, con, th)
}
\arguments{
\item{pk_pars}{Vector of pharmacokinetic parameters of length 4: (v_1, k_10, k_12, k_21)}

\item{ivt}{List with containing start of infusion times, end of infusion times,
and rate of infusion at each dose}

\item{tms}{Times to use in function evaluation}

\item{con}{Concentration of drug at each of the specified times}

\item{th}{Threshold value for effective treatment}
}
\value{
Fraction of time spent above the specified threshold from time of first dose through
12 hours after end of the last dose
}
\description{
Time Above Pharmacokinetic Threshold
}
