% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_solution.R
\name{pk_solution}
\alias{pk_solution}
\title{PK solution over dosing cycle}
\usage{
pk_solution(k_10, k_12, k_21, v_1, ivt, init = c(0, 0))
}
\arguments{
\item{k_10}{Elimination rate from central compartment}

\item{k_12}{Elimination rate from central to peripheral compartment}

\item{k_21}{Elimination rate from peripheral to central compartment}

\item{v_1}{Volume of central compartment}

\item{ivt}{List with containing start of infusion times (h), end of infusion times (h),
and rate of infusion (g/h) at each dose}

\item{init}{Initial concentration of drug in c(central, peripheral) compartments}
}
\value{
A function that will compute concentration of drug in the body given time since
first infusion.
}
\description{
The 'pk_solution' function impelments a piecewise solution
to the two-compartment model, taking into account the sequence
of dosing events:
}
