% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkm.R
\name{pkm}
\alias{pkm}
\title{Posterior concentration estimates}
\usage{
pkm(pars = c(lv_1 = 3.223, lk_10 = -1.65, lk_12 = -7, lk_21 = -7, lerr =
  2.33), ivt, dat, alp = 0.05, cod = 12, thres = 64)
}
\arguments{
\item{ivt}{List with containing start of infusion times, end of infusion times,
and rate of infusion at each dose}

\item{dat}{Concentration data frame of the form: data.frame(time_h, conc_mg_dl)}

\item{alp}{Value of alpha to use for generating pointwise 1 - alpha confidence bands}

\item{cod}{Length of time after end of last dose to consider}

\item{thres}{Threshold for effective treatment}

\item{est}{Estimate obtained from \code{optim} function with \code{Hessian = TRUE}}
}
\value{
posterior estimates
}
\description{
Posterior concentration estimates
}
\examples{

}
