% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pkm.R
\name{plot.pkm}
\alias{plot.pkm}
\title{Plot method for PK model}
\usage{
\method{plot}{pkm}(object, alp = 0.05, ...)
}
\arguments{
\item{object}{An object of class \code{pkm}}

\item{alp}{Value used to produce (1 - \code{alp})\% credible intervals}

\item{...}{Additional arguments}
}
\value{
Concentration-time curve for the fitted pharmacokinetic model
}
\description{
Plot method for PK model
}
