% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.pkm.R
\name{print.pkm}
\alias{print.pkm}
\title{Print method for pkm object}
\usage{
\method{print}{pkm}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{pkm model}

\item{digits}{number of digits to round to}

\item{...}{other parameters}
}
\value{
Printed model
}
\description{
Print method for pkm object
}
