% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_pkm.R
\name{shiny_pkm}
\alias{shiny_pkm}
\title{Shiny application for pharmacokinetic prediction model}
\usage{
shiny_pkm()
}
\value{
Launches an interaction interface for implementing Bayesian pharmacokinetic modeling

Verify there aren't any problems running this - might need to import shiny, shinydashboard, and rhandsontable libraries
}
\description{
Leave as-is? Or should any values be considered parameters
Currently specific to piperacillin (including some associated text)
}
