% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conversion.R
\name{unit_conversion}
\alias{unit_conversion}
\title{Unit conversion}
\usage{
unit_conversion(x, mvunits, forward = TRUE)
}
\arguments{
\item{x}{The value of the number to be converted, or to convert 2 (depending on `forward` argument)}

\item{mvunits}{Units of the mass and volume of the data. Must be entered as "mass/volume"}

\item{forward}{Boolean. Tells whether or not the conversion is forward (into g/L) or backward (from g/L to original)}
}
\value{
posterior estimates
}
\description{
Converts units for compatibility with package functions.
}
