% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.pkm.R
\name{update.pkm}
\alias{update.pkm}
\title{Update method for PK model}
\usage{
\method{update}{pkm}(pkm)
}
\arguments{
\item{pkm}{An object of class "pkm}
}
\description{
Update method for PK model
}
\examples{

}
