% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.mic.R
\name{confint.mic}
\alias{confint.mic}
\title{Confidence intervals for ftMIC statistic}
\usage{
\method{confint}{mic}(object, parm, level, ...)
}
\arguments{
\item{object}{"mic" object obtained from \code{mic_stat}}

\item{parm}{which parameters should be given confidence intervals}

\item{level}{Desired confidence level of the interval}

\item{...}{other arguments}
}
\value{
CI
}
\description{
Confidence intervals for ftMIC statistic
}
