% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.mic.R
\name{confint.mic}
\alias{confint.mic}
\title{Confidence intervals for ftMIC statistic}
\usage{
\method{confint}{mic}(ftmic, conf.level = 0.95)
}
\arguments{
\item{ftmic}{"mic" object obtained from \code{mic_stat}}

\item{conf.level}{Desired confidence level of the interval}
}
\value{
CI
}
\description{
Confidence intervals for ftMIC statistic
}
