% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood.R
\name{log_likelihood}
\alias{log_likelihood}
\title{Log-likelihood for PK parameters}
\usage{
log_likelihood(lpr, ivt, dat, ini = c(0, 0))
}
\arguments{
\item{lpr}{log-PK parameter vector}

\item{ivt}{List with containing start of infusion times, end of infusion times,
and rate of infusion at each dose}

\item{dat}{Concentration data frame of the form: data.frame(time_h, conc_mg_dl)}

\item{ini}{Initial concentrations in each compartment}
}
\value{
The value of the log-likelihood evaluated at the specified log-parameter vector
}
\description{
Log-likelihood for PK parameters
}
