% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_posterior.R
\name{log_posterior}
\alias{log_posterior}
\title{Log-posterior for PK parameters}
\usage{
log_posterior(
  lpr,
  ivt,
  dat,
  mu = getOption("pkpredict.pip.default.prior")$log_pk_mean,
  sig = getOption("pkpredict.pip.default.prior")$log_pk_vcov,
  ler_mean = getOption("pkpredict.pip.default.prior")$log_err_mean,
  ler_sdev = getOption("pkpredict.pip.default.prior")$log_err_sd
)
}
\arguments{
\item{lpr}{log-PK parameter vector with error: (lv_1, lk_10, lk_12, lk_21, ler_mean)}

\item{ivt}{List with containing start of infusion times (h), end of infusion times (h),
and rate of infusion (g/h) at each dose}

\item{dat}{Concentration data frame of the form: data.frame(time_h, conc_mg_dl)}

\item{mu}{prior pk param mean}

\item{sig}{prior pk vcov matrix}

\item{ler_mean}{prior error mean}

\item{ler_sdev}{prior error sd}
}
\value{
The log-posterior distribution evaluated at the specified log-parameter vector
}
\description{
Evaluates the prior distribution (normal) of the log parameter vector with
normally distributed error.
}
\examples{
ivt_d <- list(list(begin=0.0, end=0.5, k_R=6),
              list(begin=8.0, end=8.5, k_R=6),
              list(begin=16.0, end=16.5, k_R=6))
dat_d <- data.frame(time_h = c(1,4,40), conc_mcg_ml = c(82.7,80.4,60))

log_posterior(lpr = c(getOption("pkpredict.pip.default.prior")$log_pk_mean,
                      getOption("pkpredict.pip.default.prior")$log_err_mean),
              ivt = ivt_d, dat = dat_d)

}
