% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_posterior.R
\name{log_posterior}
\alias{log_posterior}
\title{Log-posterior for PK parameters}
\usage{
log_posterior(lpr, ivt, dat, mu = c(lv_1 = 3.223, lk_10 = -1.65, lk_12 = -7,
  lk_21 = -7), sig = 300 * matrix(c(0.00167, -0.00128, 0, 0, -0.00128,
  0.00154, 0, 0, 0, 0, 0.00015, 0, 0, 0, 0, 0.00015), 4, 4), ler_mean = 2.33,
  ler_sdev = 0.32)
}
\arguments{
\item{lpr}{log-PK parameter vector with error: (lv_1, lk_10, lk_12, lk_21, ler_mean)}

\item{ivt}{List with containing start of infusion times (h), end of infusion times (h),
and rate of infusion (g/h) at each dose}

\item{dat}{Concentration data frame of the form: data.frame(time_h, conc_mg_dl)}
}
\value{
The log-posterior distribution evaluated at the specified log-parameter vector
}
\description{
Evaluates the prior distribution (normal) of the log parameter vector with
normally distributed error.
}
