% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_prior.R
\name{log_prior}
\alias{log_prior}
\title{Log-prior for pharmacokinetic parameters}
\usage{
log_prior(
  lpr,
  mu = getOption("pkpredict.pip.default.prior")$log_pk_mean,
  sig = getOption("pkpredict.pip.default.prior")$log_pk_vcov,
  ler_mean = getOption("pkpredict.pip.default.prior")$log_err_mean,
  ler_sdev = getOption("pkpredict.pip.default.prior")$log_err_sd
)
}
\arguments{
\item{lpr}{log-PK parameter vector with error: (lv_1, lk_10, lk_12, lk_21, ler_mean)}

\item{mu}{log-mean of the PK parameter distribution}

\item{sig}{log-variance-covariance matrix of the PK parameter distribution}

\item{ler_mean}{log-mean of the error distribution}

\item{ler_sdev}{log-standard deviation of the error distribution}
}
\value{
The value of the log-prior evaluated at the specified log-parameter vector
}
\description{
Evaluates the prior distribution (normal) of the log parameter vector with
normally distributed error.
}
\details{
Parameter distributions:
log(lv_1, lk_10, lk_12, lk_21) ~ N(mu, sig)
log(error) ~ N(ler_mean, ler_sdev)
}
\examples{
log_prior(lpr = c(getOption("pkpredict.pip.default.prior")$log_pk_mean,
                  getOption("pkpredict.pip.default.prior")$log_err_mean))

}
