% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metro.R
\name{metropolis}
\alias{metropolis}
\title{Functions needed for MH and MCMC, currently NOT exported.}
\usage{
metropolis(theta, ivt, dat, Sigma)
}
\description{
Functions needed for MH and MCMC, currently NOT exported.
}
\examples{


}
