% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metro.R
\name{metropolis}
\alias{metropolis}
\title{Functions needed for MH and MCMC, currently NOT exported.}
\usage{
metropolis(theta, ivt, dat, Sigma)
}
\arguments{
\item{theta}{pk parameters}

\item{ivt}{infusion schedule}

\item{dat}{data}

\item{Sigma}{covariance matrix}
}
\description{
Functions needed for MH and MCMC, currently NOT exported.
}
