% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mic_stat.R
\name{mic_stat}
\alias{mic_stat}
\title{Time Above Pharmacokinetic Threshold}
\usage{
mic_stat(pars = c(lv_1 = 3.223, lk_10 = -1.65, lk_12 = -7, lk_21 = -7, lerr =
  2.33), ivt, dat, th, times = NULL, cod = 12, con = NULL, init = c(0,
  0))
}
\arguments{
\item{pars}{Vector of pharmacokinetic parameters of length 5: (log(v_1), log(k_10), log(k_12), log(k_21), log(err))}

\item{ivt}{List with containing start of infusion times (h), end of infusion times (h),
and rate of infusion (g/h) at each dose}

\item{dat}{Concentration data frame of the form: data.frame(time_h, conc_mcg_ml)}

\item{th}{Threshold value for effective treatment - check units}

\item{cod}{Length of time after end of last dose to consider}

\item{con}{Concentration of drug at each of the specified times}

\item{init}{Initial concentrations in c(central, peripheral) compartments}

\item{tms}{Times to use in function evaluation}
}
\value{
Fraction of time spent above the specified threshold from time of first dose through
\code{cod} hours after end of the last dose
}
\description{
Compute an estimate of pharmacodynamic target attainment, typically defined
to be some multiple of the minimum inhibitory concentration of the infecting microorganism
}
\examples{


}
