% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mic_stat.R
\name{mic_stat}
\alias{mic_stat}
\title{Time Above Pharmacokinetic Threshold}
\usage{
mic_stat(
  ivt,
  th,
  dat = data.frame(),
  pars = c(getOption("pkpredict.pip.default.prior")$log_pk_mean,
    getOption("pkpredict.pip.default.prior")$log_err_mean),
  cod = 12,
  timeint = NULL,
  conf.level = 0.95,
  mcmc = FALSE,
  nreps = 5000,
  nburnin = 2000,
  nthin = 10,
  seed = NULL,
  shiny = FALSE,
  ...
)
}
\arguments{
\item{ivt}{List with containing start of infusion times (h), end of infusion times (h),
and rate of infusion (g/h) at each dose. First infusion within \code{ivt} must be a true initial infusion,
i.e. concentration of the drug in the patient's system must be 0 prior to that time.}

\item{th}{Threshold value for effective treatment - check units}

\item{dat}{data}

\item{pars}{Vector of pharmacokinetic parameters of length 5: (log(v_1), log(k_10), log(k_12), log(k_21), log(err))}

\item{cod}{Length of time after end of last dose to consider}

\item{timeint}{Vector with two elements indicating the start and end of the time interval over which
the statistic is computed. By default covers the full dosing period + \code{cod}}

\item{conf.level}{Desired confidence level of the interval}

\item{mcmc}{logical: should estimate of time above threshold be computed using MCMC (false = laplace approximation)}

\item{nreps}{mcmc replications}

\item{nburnin}{mcmc burn in iterations}

\item{nthin}{mcmc thinning interval}

\item{seed}{seed for replication}

\item{shiny}{is shiny being used}

\item{...}{additional arguments (e.g., `mu`, `sig`, `ler_mean`, `ler_sdev` for changing the PK parameter prior mean,
variance-covariance matrix and error prior mean and standard deviation, respectively)}
}
\value{
Fraction of time spent above the specified threshold within a certain time interval (by default: time of first dose through
\code{cod} hours after end of the last dose).
}
\description{
Compute an estimate of pharmacodynamic target attainment, typically defined
to be some multiple of the minimum inhibitory concentration of the infecting microorganism.
}
