% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mic_stat.R
\name{mic_stat}
\alias{mic_stat}
\title{Time Above Pharmacokinetic Threshold}
\usage{
mic_stat(ivt, th, pars = c(lv_1 = 3.223, lk_10 = -1.65, lk_12 = -7, lk_21 =
  -7, lerr = 2.33), timeint = c(0, max(sapply(ivt, function(x) x$end)) + cod),
  cod = 12)
}
\arguments{
\item{ivt}{List with containing start of infusion times (h), end of infusion times (h),
and rate of infusion (g/h) at each dose}

\item{th}{Threshold value for effective treatment - check units}

\item{pars}{Vector of pharmacokinetic parameters of length 5: (log(v_1), log(k_10), log(k_12), log(k_21), log(err))}

\item{timeint}{Vector with two elements indicating the start and end of the time interval over which
the statistic is computed}

\item{cod}{Length of time after end of last dose to consider}
}
\value{
Fraction of time spent above the specified threshold from time of first dose through
\code{cod} hours after end of the last dose
}
\description{
Compute an estimate of pharmacodynamic target attainment, typically defined
to be some multiple of the minimum inhibitory concentration of the infecting microorganism
}
\examples{


}
