% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkm.R
\name{pkm}
\alias{pkm}
\title{Posterior concentration estimates}
\usage{
pkm(formula, data, subset, ivt, pars = c(lv_1 = 3.223, lk_10 = -1.65, lk_12 =
  -7, lk_21 = -7, lerr = 2.33), alp = 0.05, cod = 12, thres = 64,
  timeint = c(0, max(sapply(ivt, function(x) x$end)) + cod), mcmc = FALSE,
  nreps = 6000, nburnin = 3000, seed = NULL, shiny = FALSE, ...)
}
\arguments{
\item{formula}{A formula where the left side is the measured concentration of drug
and the right side is the times of concentration measurements}

\item{data}{Data frame with concentration data (time of measurement in hours and concentration in mcg/ml)}

\item{subset}{Subset of the \code{dat} data frame to use}

\item{ivt}{List with containing start of infusion times, end of infusion times,
and rate of infusion at each dose}

\item{pars}{Vector of (prior) log-pharmacokinetic parameters of length 5: (lv_1, lk_10, lk_12, lk_21, lerr)}

\item{alp}{Value of alpha to use for generating pointwise (1 - \code{alp})\% confidence bands}

\item{cod}{Length of time after end of last dose to consider}

\item{thres}{Threshold for effective treatment (mcg/ml)}

\item{timeint}{time interval over which to compute estimate}

\item{mcmc}{logical: should estimate of time above threshold be computed using MCMC (false = laplace approximation)}

\item{nreps}{number of MCMC iterations to perform (including burn in)}

\item{nburnin}{number of burn in replications to perform}

\item{seed}{seed for replicating MCMC results}

\item{shiny}{is this being used within shiny_pkm}

\item{...}{other parameters}
}
\value{
posterior estimates
}
\description{
Fits a two-compartment model to obtain posterior estimates of concentration of drug
over time.
}
\details{
Measurements must be entered in particular units: mcg/ml for concentrations, g/h in rate of
infusion, hours for times.
}
\examples{
# Insert example from Bayes.R
# pkm(concentration ~ time, dat_d, ivt_d) # something like this
}
