% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.pkm.R
\name{predict.pkm}
\alias{predict.pkm}
\title{Predict method for PK model}
\usage{
\method{predict}{pkm}(object, times)
}
\arguments{
\item{object}{Object of class "pkm"}

\item{times}{Times at which predictions of blood concentration measurements are desired}
}
\value{
Predicted blood concentration measurements
}
\description{
Predict method for PK model
}
\examples{

}
