% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.pkm.R
\name{print.pkm}
\alias{print.pkm}
\title{Title}
\usage{
\method{print}{pkm}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{pkm model x}

\item{digits}{number of digits to round to}

\item{...}{other parameters}
}
\value{
printed model x
}
\description{
Title
}
