% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conversion.R
\name{unit_conversion}
\alias{unit_conversion}
\title{Unit conversion}
\usage{
unit_conversion(x, mvunits, to_gL = TRUE)
}
\arguments{
\item{x}{The value of the number to be converted, or to convert to (depending on `to_gL` argument)}

\item{mvunits}{Units of the mass and volume of the data. Must be entered as "mass/volume"}

\item{to_gL}{Boolean. Tells whether or not the conversion is into g/L or bfrom g/L to original}
}
\value{
Value equivalent to input with the appropriate units.
}
\description{
Converts units for compatibility with package functions.
}
\examples{
unit_conversion(64, "mcg/mL")
}
